Splitted files and removed `IMPALA_CLIENT` object
