+++ src/ridi/impala/config.py
@@ -7,14 +7,14 @@ ConnectArgsType = namedtuple('ConnectArgsType', [
 IMPALA_CONNECT_ARGS = None
 
 
-class ImpalaConfig(object):
-    @staticmethod
-    def set(host=None, port=None, user=None, password=None, database=None):
-        global IMPALA_CONNECT_ARGS
-        IMPALA_CONNECT_ARGS = ConnectArgsType(
-            host=host,
-            port=port,
-            user=user,
-            password=password,
-            database=database,
-        )._asdict()
+def configure(host=None, port=None, user=None, password=None, database=None):
+    global IMPALA_CONNECT_ARGS
+    IMPALA_CONNECT_ARGS = ConnectArgsType(
+        host=host,
+        port=port,
+        user=user,
+        password=password,
+        database=database,
+    )._asdict()
+
+__all__ = ["configure"]
